Fix narrowing conversion error in scaleDpi function
Change std::lround return type from long to int using static_cast
to resolve clang-tidy warnings:
- bugprone-narrowing-conversions
- clang-diagnostic-shorten-64-to-32

+++ src/common/gui/theme.h
@@ -30,7 +30,7 @@ inline qreal getDpiScaleFactor() {
 
 /// Scale a value based on DPI
 inline int scaleDpi(int baseValue) {
-  return std::lround(baseValue * getDpiScaleFactor());
+  return static_cast<int>(std::lround(baseValue * getDpiScaleFactor()));
 }
 
 /// Color palette as defined in client_ui_design.md
